+++ src/scores/score.ts
@@ -91,6 +91,34 @@ export default class Score extends CollectionWithMetadata<Melody> {
     }
 
     /**
+     * Return a new Score where all notes with identical duration, start tick
+     * and volume are combined into a single chord.
+     * 
+     * This is an experimental feature: the API should *not* be considered stable.
+     *
+     * Caveats:
+     *  - all tracks are collapsed into a single track
+     *  - metadata associated with tracks beyond the first are lost
+     *  - metadata associated with notes beyond the first in a chord is lost
+     *  - note order is not retained
+     *  - exact ticks are applied to all notes within the Score
+     *  - notes in different instruments may be collapsed into one chord
+     */
+    withChordsCombined(): this {
+        if (this.length === 0) { return this; }
+
+        const exact = this.withAllTicksExact();
+        const mel = exact.contents[0].append(...exact.contents.slice(1))
+            .sort((a, b) => ((a.at as number) - (b.at as number)) || (a.duration - b.duration))
+            .replaceIfWindow(2, 1,
+                ([ curr, next ]) => (curr.at as number) === (next.at as number) && curr.duration === next.duration && curr.velocity === next.velocity,
+                ([ curr, next ]) => curr.setPitches([ ...curr.pitch.pitches(), ...next.pitch.pitches() ])
+            );
+
+        return this.construct([ mel ]);
+    }
+
+    /**
      * Returns an HTML canvas visualisation of the Score. Will attempt to size the canvas
      * as close as possible to the sizes supplied.
      *